Changes to allow testPyfitsImage.py run on both 32 and 64 bit machines
git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@354 ed100bfc-0583-0410-97f2-c26b58777a21

+++ test/testPyfitsImage.py
@@ -404,7 +404,7 @@ CONTINUE  '&' / comments in line 1 comments with ''.                            
         # create an HDU with header and data
         # notice that the header has the right NAXIS's since it is constructed with
         # ImageHDU
-        hdu2=pyfits.ImageHDU(header=r[1].header, data=numpy.array([1,2]))
+        hdu2=pyfits.ImageHDU(header=r[1].header, data=numpy.array([1,2],dtype='int32'))
         
         tmpfile = open(jfile,'w')
         sys.stdout = tmpfile
